requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
djangorestframework==3.11.0
environs==7.2.0
gunicorn==20.0.4
idna==2.9
marshmallow==3.5.0
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
tinkoff-invest-openapi-client==0.0.6
urllib3==1.25.8
websocket-client==0.56.0
